Repackages CLI, consolidates PicoCLI commands, and addresses comments
