+++ partiql-app/partiql-cli/src/main/kotlin/org/partiql/cli/format/SexpFormatter.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2022 Amazon.com, Inc. or its affiliates.  All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at:
+ *
+ *      http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+ * language governing permissions and limitations under the License.
+ */
+
+package org.partiql.cli.format
+
+import com.amazon.ion.system.IonTextWriterBuilder
+import org.partiql.pig.runtime.DomainNode
+
+internal object SexpFormatter : NodeFormatter {
+    override fun format(input: DomainNode): String = buildString {
+        val writer = IonTextWriterBuilder.pretty().build(this)
+        input.toIonElement().writeTo(writer)
+    }
+}
